+++ src/include/background.h
@@ -10,6 +10,8 @@
 #define MAX_BACKGROUND_THREADS  20      /* The total number threads allowed to be run from within the MOO.
                                            Can be overridden with $server_options.max_background_threads */
 
+static threadpool background_pool;
+
 typedef struct background_waiter {
     vm the_vm;                          // Where we resume when we're done.
     int handle;                         // Our position in the process table.
@@ -19,10 +21,11 @@ typedef struct background_waiter {
     int fd[2];                          // The pipe used to resume the task immediately.
     Var return_value;                   // The final return value that gets sucked up by the network callback.
     char *human_title;                  // A human readable explanation for the thread's existance.
+    threadpool *pool;                   // The thread pool to put this into.
 } background_waiter;
 
 // User-visible functions
-extern package background_thread(void (*callback)(Var, Var*), Var* data, char *human_title);
+extern package background_thread(void (*callback)(Var, Var*), Var* data, char *human_title, threadpool *the_pool = &background_pool);
 extern bool can_create_thread();
 extern void make_error_map(enum error error_type, const char *msg, Var *ret);