+++ src/include/options.h
@@ -511,6 +511,8 @@
  * Configurable options for the background subsystem.
  * TOTAL_BACKGROUND_THREADS is the total number of pthreads that will be created
  * at runtime to process background MOO tasks.
+ * TOTAL_DNS_THREADS is the total number of pthreads that will be created at
+ * runtime to process DNS name lookups from the name_lookup() function.
  * DEFAULT_THREAD_MODE dictates the default behavior of threaded MOO functions
  * without a call to set_thread_mode. When set to true, the default behavior is
  * to thread these functions, requiring a call to set_thread_mode(0) to disable.
@@ -520,6 +522,7 @@
  */
 
 #define TOTAL_BACKGROUND_THREADS    1
+#define TOTAL_DNS_THREADS           1
 #define DEFAULT_THREAD_MODE         true
 
 /******************************************************************************